audio/src4/requirements.txt
micropython-machine
micropython-time
micropython-os
micropython-usb-audio